Remove spack.util.imp, since code is only used in spack.repo
